add excel for thietBi and xuly

+++ src/main/java/com/example/project_2/DTO/ThietBi.java
@@ -15,7 +15,7 @@ import jakarta.persistence.*;
 public class ThietBi {
     @Id
     @Column(name = "MaTB", nullable = false)
-    private Integer MaTB;
+    private long MaTB;
 
     @Column(name = "TenTB", nullable = false, length = 100)
     private String TenTB;
@@ -26,17 +26,17 @@ public class ThietBi {
     public ThietBi() {
     }
     
-    public ThietBi(int MaTB, String TenTB, String MoTaTB) {
+    public ThietBi(long MaTB, String TenTB, String MoTaTB) {
         this.MaTB = MaTB;
         this.TenTB = TenTB;
         this.MoTaTB = MoTaTB;
     }
     
-    public Integer getMaTB() {
+    public long getMaTB() {
         return MaTB;
     }
 
-    public void setMaTB(Integer MaTB) {
+    public void setMaTB(long MaTB) {
         this.MaTB = MaTB;
     }
